runtime: add pallet-session with static genesis validator set
Wires `pallet-session` into the runtime so dashboards and polkadot.js
clients can read `api.query.session.validators`. The session API also
provides the standard surface for `author_rotateKeys` once hybrid key
rotation lands.

The genesis validator set is static for v0.2: `SessionManager = ()`
returns `None` on every `new_session`, so the keys registered in the
`SessionConfig.keys` at genesis persist forever. Each genesis authority
is now a triple of `(account, babe, grandpa)`; the same account is used
as both validator stash and controller since staking is not wired in.

`ShouldEndSession`/`NextSessionRotation` are delegated to BABE so
session boundaries line up with BABE epochs. BABE keeps the existing
`SameAuthoritiesForever` trigger — sessions rotate cosmetically but the
authority set never changes.

Drive-by: drop a `u32::from(raw)` on a `u32` in `quantum-validation`'s
solution packer that clippy started flagging once the workspace built
with pallet-session pulled in.

Bumps `spec_version` 101 -> 102 so clients regenerate the metadata
exposing the new `Session` storage entries; `transaction_version`
unchanged.

+++ crates/quantum-validation/src/packed.rs
@@ -128,7 +128,7 @@ pub fn pack_solution(
         let byte_index = bit_offset / 8;
         let intra_byte = bit_offset % 8;
 
-        let shifted = u32::from(raw) << intra_byte;
+        let shifted = raw << intra_byte;
         out[byte_index] |= (shifted & 0xFF) as u8;
         if byte_index + 1 < out.len() {
             out[byte_index + 1] |= ((shifted >> 8) & 0xFF) as u8;
